Fix compilation warnings in module JCAI (gcc 4.0.2).
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@448 37977431-3df6-0310-b722-df95706aa16b

+++ JCAI/AIConfigLoader.cpp
@@ -90,7 +90,7 @@ bool CfgBuildOptions::Parse (CfgBuffer& buf)
 				if (!val.Parse (buf))
 					return false;
 
-				opt->count = val.value;
+				opt->count = (int) val.value;
 			}
 			else if (*buf == '{') // its a set of conditions
 			{
